+++ src/main/groovy/com/bmuschko/gradle/nexus/ExtraArchivePluginExtension.groovy
@@ -1,0 +1,7 @@
+package com.bmuschko.gradle.nexus
+
+class ExtraArchivePluginExtension {
+    Boolean sources = Boolean.TRUE
+    Boolean tests = Boolean.FALSE
+    Boolean javadoc = Boolean.TRUE
+}